+++ buildhat/hat.py
@@ -1,21 +1,24 @@
 """HAT handling functionality"""
 
+from concurrent.futures import Future
+
 from .devices import Device
 
 
 class Hat:
     """Allows enumeration of devices which are connected to the hat"""
 
-    def __init__(self, device=None):
+    def __init__(self, device=None, debug=False):
         """Hat
 
         :param device: Optional string containing path to Build HAT serial device
+        :param debug: Optional boolean to log debug information
         """
         self.led_status = -1
-        if not device:
-            Device._setup()
+        if device is None:
+            Device._setup(debug=debug)
         else:
-            Device._setup(device)
+            Device._setup(device=device, debug=debug)
 
     def get(self):
         """Get devices which are connected or disconnected
@@ -44,16 +47,15 @@ class Hat:
         :return: Voltage on the input power jack
         :rtype: float
         """
+        ftr = Future()
+        Device._instance.vinftr.append(ftr)
         Device._instance.write(b"vin\r")
-        with Device._instance.vincond:
-            Device._instance.vincond.wait()
-
-        return Device._instance.vin
+        return ftr.result()
 
     def _set_led(self, intmode):
         if isinstance(intmode, int) and intmode >= -1 and intmode <= 3:
             self.led_status = intmode
-            Device._instance.write("ledmode {}\r".format(intmode).encode())
+            Device._instance.write(f"ledmode {intmode}\r".encode())
 
     def set_leds(self, color="voltage"):
         """Set the two LEDs on or off on the BuildHAT.