+++ buildhat/serinterface.py
@@ -1,6 +1,8 @@
 """Build HAT handling functionality"""
 
+import logging
 import queue
+import tempfile
 import threading
 import time
 from enum import Enum
@@ -29,6 +31,8 @@ class Connection:
         self.typeid = -1
         self.connected = False
         self.callit = None
+        self.simplemode = -1
+        self.combimode = -1
 
     def update(self, typeid, connected, callit=None):
         """Update connection information for port
@@ -69,46 +73,54 @@ class BuildHAT:
     RESET_GPIO_NUMBER = 4
     BOOT0_GPIO_NUMBER = 22
 
-    def __init__(self, firmware, signature, version, device="/dev/serial0"):
+    def __init__(self, firmware, signature, version, device="/dev/serial0", debug=False):
         """Interact with Build HAT
 
         :param firmware: Firmware file
         :param signature: Signature file
         :param version: Firmware version
         :param device: Serial device to use
+        :param debug: Optional boolean to log debug information
         :raises BuildHATError: Occurs if can't find HAT
         """
         self.cond = Condition()
         self.state = HatState.OTHER
         self.connections = []
-        self.portcond = []
-        self.pulsecond = []
-        self.rampcond = []
+        self.portftr = []
+        self.pulseftr = []
+        self.rampftr = []
+        self.vinftr = []
+        self.motorqueue = []
         self.fin = False
         self.running = True
-        self.vincond = Condition()
-        self.vin = None
+        if debug:
+            tmp = tempfile.NamedTemporaryFile(suffix=".log", prefix="buildhat-", delete=False)
+            logging.basicConfig(filename=tmp.name, format='%(asctime)s %(message)s',
+                                level=logging.DEBUG)
 
         for _ in range(4):
             self.connections.append(Connection())
-            self.portcond.append(Condition())
-            self.pulsecond.append(Condition())
-            self.rampcond.append(Condition())
+            self.portftr.append([])
+            self.pulseftr.append([])
+            self.rampftr.append([])
+            self.motorqueue.append(queue.Queue())
 
         self.ser = serial.Serial(device, 115200, timeout=5)
         # Check if we're in the bootloader or the firmware
         self.write(b"version\r")
 
+        emptydata = 0
         incdata = 0
         while True:
-            try:
-                line = self.ser.readline().decode('utf-8', 'ignore')
-            except serial.SerialException:
-                pass
+            line = self.read()
             if len(line) == 0:
-                # Didn't recieve any data
-                break
-            if line[:len(BuildHAT.FIRMWARE)] == BuildHAT.FIRMWARE:
+                # Didn't receive any data
+                emptydata += 1
+                if emptydata > 3:
+                    break
+                else:
+                    continue
+            if cmp(line, BuildHAT.FIRMWARE):
                 self.state = HatState.FIRMWARE
                 ver = line[len(BuildHAT.FIRMWARE):].split(' ')
                 if int(ver[0]) == version:
@@ -117,7 +129,7 @@ class BuildHAT:
                 else:
                     self.state = HatState.NEEDNEWFIRMWARE
                     break
-            elif line[:len(BuildHAT.BOOTLOADER)] == BuildHAT.BOOTLOADER:
+            elif cmp(line, BuildHAT.BOOTLOADER):
                 self.state = HatState.BOOTLOADER
                 break
             else:
@@ -141,15 +153,22 @@ class BuildHAT:
         self.cb.daemon = True
         self.cb.start()
 
+        for q in self.motorqueue:
+            ml = threading.Thread(target=self.motorloop, args=(q,))
+            ml.daemon = True
+            ml.start()
+
         # Drop timeout value to 1s
+        listevt = threading.Event()
         self.ser.timeout = 1
-        self.th = threading.Thread(target=self.loop, args=(self.cond, self.state == HatState.FIRMWARE, self.cbqueue))
+        self.th = threading.Thread(target=self.loop, args=(self.cond, self.state == HatState.FIRMWARE, self.cbqueue, listevt))
         self.th.daemon = True
         self.th.start()
 
         if self.state == HatState.FIRMWARE:
             self.write(b"port 0 ; select ; port 1 ; select ; port 2 ; select ; port 3 ; select ; echo 0\r")
             self.write(b"list\r")
+            listevt.set()
         elif self.state == HatState.NEEDNEWFIRMWARE or self.state == HatState.BOOTLOADER:
             self.write(b"reboot\r")
 
@@ -182,17 +201,17 @@ class BuildHAT:
             sig = f.read()
         self.write(b"clear\r")
         self.getprompt()
-        self.write("load {} {}\r".format(len(firm), self.checksum(firm)).encode())
+        self.write(f"load {len(firm)} {self.checksum(firm)}\r".encode())
         time.sleep(0.1)
-        self.write(b"\x02")
-        self.write(firm)
-        self.write(b"\x03\r")
+        self.write(b"\x02", replace="0x02")
+        self.write(firm, replace="--firmware file--")
+        self.write(b"\x03\r", replace="0x03")
         self.getprompt()
-        self.write("signature {}\r".format(len(sig)).encode())
+        self.write(f"signature {len(sig)}\r".encode())
         time.sleep(0.1)
-        self.write(b"\x02")
-        self.write(sig)
-        self.write(b"\x03\r")
+        self.write(b"\x02", replace="0x02")
+        self.write(sig, replace="--signature file--")
+        self.write(b"\x03\r", replace="0x03")
         self.getprompt()
 
     def getprompt(self):
@@ -201,12 +220,8 @@ class BuildHAT:
         Need to decide what we will do, when no prompt
         """
         while True:
-            line = b""
-            try:
-                line = self.ser.readline().decode('utf-8', 'ignore')
-            except serial.SerialException:
-                pass
-            if line[:len(BuildHAT.PROMPT)] == BuildHAT.PROMPT:
+            line = self.read()
+            if cmp(line, BuildHAT.PROMPT):
                 break
 
     def checksum(self, data):
@@ -224,12 +239,33 @@ class BuildHAT:
             u = (u ^ data[i]) & 0xFFFFFFFF
         return u
 
-    def write(self, data):
+    def write(self, data, log=True, replace=""):
         """Write data to the serial port of Build HAT
 
         :param data: Data to write to Build HAT
+        :param log: Whether to log line or not
+        :param replace: Whether to log an alternative string
         """
         self.ser.write(data)
+        if not self.fin and log:
+            if replace != "":
+                logging.debug(f"> {replace}")
+            else:
+                logging.debug(f"> {data.decode('utf-8', 'ignore').strip()}")
+
+    def read(self):
+        """Read data from the serial port of Build HAT
+
+        :return: Line that has been read
+        """
+        line = ""
+        try:
+            line = self.ser.readline().decode('utf-8', 'ignore').strip()
+        except serial.SerialException:
+            pass
+        if line != "":
+            logging.debug(f"< {line}")
+        return line
 
     def shutdown(self):
         """Turn off the Build HAT devices"""
@@ -238,17 +274,34 @@ class BuildHAT:
             self.running = False
             self.th.join()
             self.cbqueue.put(())
+            for q in self.motorqueue:
+                q.put((None, None))
             self.cb.join()
             turnoff = ""
             for p in range(4):
                 conn = self.connections[p]
                 if conn.typeid != 64:
-                    turnoff += "port {} ; pwm ; coast ; off ;".format(p)
+                    turnoff += f"port {p} ; pwm ; coast ; off ;"
                 else:
-                    self.write("port {} ; write1 {}\r".format(p, ' '.join('{:x}'.format(h)
-                                                              for h in [0xc2, 0, 0, 0, 0, 0, 0, 0, 0, 0])).encode())
-            self.write("{}\r".format(turnoff).encode())
+                    hexstr = ' '.join(f'{h:x}' for h in [0xc2, 0, 0, 0, 0, 0, 0, 0, 0, 0])
+                    self.write(f"port {p} ; write1 {hexstr}\r".encode())
+            self.write(f"{turnoff}\r".encode())
             self.write(b"port 0 ; select ; port 1 ; select ; port 2 ; select ; port 3 ; select ; echo 0\r")
+
+    def motorloop(self, q):
+        """Event handling for non-blocking motor commands
+
+        :param q: Queue of motor functions
+        """
+        while self.running:
+            func, data = q.get()
+            if func is None:
+                break
+            else:
+                func(*data)
+                func = None  # Necessary for 'del' to function correctly on motor object
+                data = None
+                q.task_done()
 
     def callbackloop(self, q):
         """Event handling for callbacks
@@ -266,7 +319,7 @@ class BuildHAT:
             cb[0]()(cb[1])
             q.task_done()
 
-    def loop(self, cond, uselist, q):
+    def loop(self, cond, uselist, q, listevt):
         """Event handling for Build HAT
 
         :param cond: Condition used to block user's script till we're ready
@@ -275,11 +328,7 @@ class BuildHAT:
         """
         count = 0
         while self.running:
-            line = b""
-            try:
-                line = self.ser.readline().decode('utf-8', 'ignore')
-            except serial.SerialException:
-                pass
+            line = self.read()
             if len(line) == 0:
                 continue
             if line[0] == "P" and line[2] == ":":
@@ -289,13 +338,13 @@ class BuildHAT:
                     typeid = int(line[2 + len(BuildHAT.CONNECTED):], 16)
                     self.connections[portid].update(typeid, True)
                     if typeid == 64:
-                        self.write("port {} ; on\r".format(portid).encode())
-                    if uselist:
+                        self.write(f"port {portid} ; on\r".encode())
+                    if uselist and listevt.is_set():
                         count += 1
                 elif cmp(msg, BuildHAT.CONNECTEDPASSIVE):
                     typeid = int(line[2 + len(BuildHAT.CONNECTEDPASSIVE):], 16)
                     self.connections[portid].update(typeid, True)
-                    if uselist:
+                    if uselist and listevt.is_set():
                         count += 1
                 elif cmp(msg, BuildHAT.DISCONNECTED):
                     self.connections[portid].update(-1, False)
@@ -303,14 +352,14 @@ class BuildHAT:
                     self.connections[portid].update(-1, False)
                 elif cmp(msg, BuildHAT.NOTCONNECTED):
                     self.connections[portid].update(-1, False)
-                    if uselist:
+                    if uselist and listevt.is_set():
                         count += 1
                 elif cmp(msg, BuildHAT.RAMPDONE):
-                    with self.rampcond[portid]:
-                        self.rampcond[portid].notify()
+                    ftr = self.rampftr[portid].pop()
+                    ftr.set_result(True)
                 elif cmp(msg, BuildHAT.PULSEDONE):
-                    with self.pulsecond[portid]:
-                        self.pulsecond[portid].notify()
+                    ftr = self.pulseftr[portid].pop()
+                    ftr.set_result(True)
 
             if uselist and count == 4:
                 with cond:
@@ -326,7 +375,7 @@ class BuildHAT:
 
             if line[0] == "P" and (line[2] == "C" or line[2] == "M"):
                 portid = int(line[1])
-                data = line[5:].strip().split(" ")
+                data = line[5:].split(" ")
                 newdata = []
                 for d in data:
                     if "." in d:
@@ -334,15 +383,22 @@ class BuildHAT:
                     else:
                         if d != "":
                             newdata.append(int(d))
+                # Check data was for our current mode
+                if line[2] == "M" and self.connections[portid].simplemode != int(line[3]):
+                    continue
+                elif line[2] == "C" and self.connections[portid].combimode != int(line[3]):
+                    continue
                 callit = self.connections[portid].callit
                 if callit is not None:
                     q.put((callit, newdata))
                 self.connections[portid].data = newdata
-                with self.portcond[portid]:
-                    self.portcond[portid].notify()
+                try:
+                    ftr = self.portftr[portid].pop()
+                    ftr.set_result(newdata)
+                except IndexError:
+                    pass
 
-            if len(line) >= 5 and line[1] == "." and line.strip().endswith(" V"):
-                vin = float(line.strip().split(" ")[0])
-                self.vin = vin
-                with self.vincond:
-                    self.vincond.notify()
+            if len(line) >= 5 and line[1] == "." and line.endswith(" V"):
+                vin = float(line.split(" ")[0])
+                ftr = self.vinftr.pop()
+                ftr.set_result(vin)